pyproject.toml
# ===================================================================
# METADADOS E DEPENDÊNCIAS DO PROJETO
# ===================================================================
[project]
name = "study-langchain"
version = "0.1.0"
description = "Estudos sobre langchain"
readme = "README.md"
authors = [{ name = "Fabio Lima", email = "lima.fisico@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[openai]>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-openai>=0.3.27",
    "langchain-postgres>=0.0.15",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.4",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "langchain-chroma>=0.2.0",
    "beautifulsoup4>=4.12.2",
    "pypdf>=5.1.0",
    "psycopg[binary]>=3.2.4",
    "setuptools>=75.8.0",
    "langsmith>=0.3.2",
    "langgraph-checkpoint-sqlite>=2.0.3",
    "duckduckgo-search>=7.3.0",
    "langgraph-cli>=0.1.73",
    "grandalf>=0.8",
    "ruff>=0.12.4",
    "mypy>=1.17.1",
    "pydantic-settings>=2.10.1",
    "loguru>=0.7.3",
    "ruff-lsp>=0.0.62",
    "pandas>=2.3.2",
]

[project.optional-dependencies]
dev = ["ruff", "pyright", "pytest", "pytest-xdist"]


# ===================================================================
# LINT E FORMATAÇÃO (RUFF)
# ===================================================================
[tool.ruff]
# Configurações globais que se aplicam a tudo.
line-length = 88
indent-width = 4
exclude = ["venv", ".venv", "env", ".env", "node_modules", "__pycache__"]
target-version = "py313"

# CORREÇÃO: 'fix' e 'show-fixes' são configurações GLOBAIS e ficam aqui.
fix = true
show-fixes = true

# --- Seção específica para o LINTER ---
[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "CPY001", "T201", "COM812", "D100", "D104", "D107", "D203", "D213",
]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN201", "S101"]

[tool.ruff.lint.isort]
known-first-party = ["study_langchain", "movie_project"]

# --- Seção específica para o FORMATADOR ---
[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"


# ===================================================================
# TIPAGEM (PYRIGHT)
# ===================================================================
[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.13"
include = ["src", "tests"]
exclude = [
    "**/venv", "**/.venv", "**/env", "**/.env", "**/node_modules", "**/__pycache__",
]
venv = ".venv"
venvPath = "."
executionEnvironments = [{ root = "src" }]


# ===================================================================
# TESTES (PYTEST)
# ===================================================================
[tool.pytest.ini_options]
addopts = "-s --color=yes --tb=short"
pythonpath = ["src"]
testpaths = ["tests"]
